phoromatic: Poweroff tweaking

+++ pts-core/modules/phoromatic.php
@@ -742,7 +742,7 @@ class phoromatic extends pts_module_interface
 
 						echo PHP_EOL . 'Phoromatic received a remote command to shutdown.' . PHP_EOL;
 						phoromatic::update_system_status('Attempting System Shutdown');
-						if(pts_client::executable_in_path('systemctl'))
+						if(pts_client::executable_in_path('systemctl') && rand(0, 1) == 1) // some systems like systemctl poweroff, others just like poweroff, but not consistent one method for all systems in testing
 						{
 							// Try systemd's poweroff method first
 							shell_exec('systemctl poweroff');
